% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/format_citation.R
\name{add_line_breaks}
\alias{add_line_breaks}
\title{Add line breaks to one or more strings}
\usage{
add_line_breaks(data, n = 50)
}
\arguments{
\item{data}{Either a string or a vector; if the vector is not of class character if will be coerced to one using as.character.}

\item{n}{Numeric: The number of characters that should separate consecutive line breaks.}
}
\value{
Returns the input vector unaltered except for the addition of line breaks.
}
\description{
This function takes a vector of strings and adds line breaks every n characters. Primarily built to be called internally by format_citation, this function has been made available as it can be useful in other contexts.
}
\details{
Line breaks are only added between words, so the value of n is acutally a threshold value rather than being matched exactly.
}
\examples{
add_line_breaks(data=c("On the Origin of Species"), n=2)
}
