% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/bibliography-methods.R
\name{summary.bibliography}
\alias{summary.bibliography}
\alias{print.bibliography}
\alias{[.bibliography}
\alias{c.bibliography}
\alias{as.data.frame.bibliography}
\alias{as.bibliography}
\alias{summary.bibliography,}
\alias{print.bibliography,}
\alias{c.bibliography,}
\title{Methods for class 'bibliography'}
\usage{
\method{summary}{bibliography}(object, ...)

\method{print}{bibliography}(x, n, ...)

\method{[}{bibliography}(x, n)

\method{c}{bibliography}(...)

\method{as.data.frame}{bibliography}(x, ...)

as.bibliography(x, ...)
}
\arguments{
\item{object}{An object of class 'bibliography'}

\item{...}{Any further information}

\item{x}{An object of class 'bibliography'}

\item{n}{Number of items to select/print}
}
\description{
This is a small number of standard methods for interacting with class 'bibliography'. More may be added later.
}
\section{Methods (by generic)}{
\itemize{
\item \code{summary}: Summarize a bibliography

\item \code{print}: Print a bibliography

\item \code{[}: Class bibliography

\item \code{c}: Class bibliography

\item \code{as.data.frame}: Convert a data.frame to a bibliography
}}

