% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/bibliography.R, R/bibliography_functions.R
\name{bibliography-class}
\alias{bibliography-class}
\alias{summary.bibliography,}
\alias{print.bibliography,}
\alias{c.bibliography,}
\alias{as.data.frame.bibliography}
\alias{summary.bibliography}
\alias{print.bibliography}
\alias{[.bibliography}
\alias{c.bibliography}
\alias{as.bibliography}
\title{bibliography-class}
\usage{
\method{summary}{bibliography}(object, ...)

\method{print}{bibliography}(x, n, ...)

\method{[}{bibliography}(x, n)

\method{c}{bibliography}(...)

\method{as.data.frame}{bibliography}(x, ...)

as.bibliography(x, ...)
}
\arguments{
\item{object}{An object of class 'bibliography'}

\item{...}{Any further information}

\item{x}{An object of class 'bibliography'}

\item{n}{Number of items to select/print}
}
\description{
This is a small number of standard methods for interacting with class 'bibliography'. More may be added later.
}
\details{
Methods for class bibliography
}
