% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/import_refs.R
\name{bibvris}
\alias{bibvris}
\title{Detects if a file is bib or ris}
\usage{
bibvris(z)
}
\arguments{
\item{z}{a character vector}
}
\value{
a file type: either bib, ris, or unknown
}
\description{
given a character vector, determines if the underlying file is more bib-like or ris-like
}
