% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/deduplication_functions.R
\name{calculate_similarity}
\alias{calculate_similarity}
\title{Calculates similarity between documents}
\usage{
calculate_similarity(dfm)
}
\arguments{
\item{dfm}{a document-feature matrix}
}
\value{
a data frame with text distance scores for paired documents
}
\description{
Computes the distance between documents in a document-feature matrix based on shared word associations.
}
