% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/import_functions.R
\name{check_filetypes}
\alias{check_filetypes}
\title{Checks directory files to see if they can be imported}
\usage{
check_filetypes(import_files)
}
\arguments{
\item{import_files}{a character vector with paths to files for import}
}
\value{
a character vector with paths to files for import with unsupported filetypes removed
}
\description{
Checks files staged for import to ensure they are a matching filetype and removes unsupported filetypes or directories from the list.
}
