% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/deduplication_functions.R
\name{chimera_detect}
\alias{chimera_detect}
\title{Detect multilingual texts}
\usage{
chimera_detect(text, overlap = 0.5)
}
\arguments{
\item{text}{the text to check for chimeras}

\item{overlap}{how large of a window to check for chimeras; an overlap of .5 will split the text into two chunks split at the middle, whereas higher levels of overlap will result in larger chunks to check}
}
\value{
if a chimera is detected, the language codes of contained languages
}
\description{
Detects texts that contain two languages in discrete blocks, such as a translation of a title
}
