% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/clean_functions.R
\name{clean_}
\alias{clean_}
\alias{clean_df}
\alias{clean_authors}
\alias{clean_colnames}
\title{Clean a \code{tibble} or vector}
\usage{
clean_df(x)

clean_authors(x)

clean_colnames(x)
}
\arguments{
\item{x}{A \code{tibble} with bibliographic information.}
}
\value{
Returns the input, but cleaner.
}
\description{
Cleans column and author names
}
\examples{
df <-  data.frame(
  X..title. = c(
    "EviAtlas: a tool for visualising evidence synthesis databases",
    "revtools: An R package to support article screening for evidence synthesis",
    "An automated approach to identifying search terms for systematic reviews",
    "Reproducible, flexible and high-throughput data extraction from primary literature"),
  YEAR = c("2019", "2019", "2019", "2019"),
  authors = c(
    "Haddaway et al",
    "Westgate",
    "EM Grames AND AN Stillman  & MW Tingley and CS Elphick",
    "Pick et al"))
clean_df(df)

# or use sub-functions
clean_colnames(df)
clean_authors(df)

}
