% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/clean_df.R
\name{clean_author_delimiters}
\alias{clean_author_delimiters}
\title{Detect author delimiters}
\usage{
clean_author_delimiters(x)
}
\arguments{
\item{x}{a character vector of author names}
}
\value{
a cleaned up character vector of author names
}
\description{
Standardizes delimiters between author names.
}
