% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/clean_df.R
\name{clean_authors}
\alias{clean_authors}
\title{Standardize author delimiters}
\usage{
clean_authors(x)
}
\arguments{
\item{x}{Either a string or a vector of author names.}
}
\value{
Returns the input vector with standardized delimiters.
}
\description{
This function standardizes delimiters between author names.
}
\examples{
clean_authors(c("Darwin AND Wallace",  "Darwin & Wallace"))
}
