% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/clean_df.R
\name{clean_df}
\alias{clean_df}
\title{Cleans data.frames into revtools format}
\usage{
clean_df(data)
}
\arguments{
\item{data}{a data.frame with bibliographic information}
}
\value{
a data.frame
}
\description{
Cleans column names and labels in data.frames into a format suitable for revtools applications.
}
