% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/clean_df.R
\name{clean_names}
\alias{clean_names}
\title{Clean common issues with column names}
\usage{
clean_names(x)
}
\arguments{
\item{x}{a vector of column names}
}
\value{
a cleaned vector of column names
}
\description{
Cleans messy column names from .csv imports such as extra punctuation.
}
