% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/data.R
\docType{data}
\name{code_lookup}
\alias{code_lookup}
\title{Bibliographic code lookup for search results assembly}
\format{A data frame with 207 obs of 3 variables

\describe{
 \item{code}{code used in search results}
 \item{description}{description of field from original source}
 \item{field}{bibliographic field that codes correspond to}
 \item{order}{the order in which to rank fields in assembled results}}}
\usage{
code_lookup
}
\description{
A data frame containing that can be used to look up
common codes for different bibliographic fields
across databases and merge them to a common format
}
\keyword{datasets}
