% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/deduplication_functions.R
\name{create_dfm}
\alias{create_dfm}
\title{Create a document-feature matrix}
\usage{
create_dfm(elements, language)
}
\arguments{
\item{elements}{a character vector of document information (e.g. document titles or abstracts)}

\item{language}{the language to use for tokenizing documents}
}
\value{
a matrix with documents as rows and terms as columns
}
\description{
Given a character vector of document information and a language, creates a document-feature matrix.
}
