% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/generic_text_functions.R
\name{create_dtm}
\alias{create_dtm}
\title{Construct a document-term matrix (DTM)}
\usage{
create_dtm(
  x,
  stop_words,
  bigram_check = TRUE,
  bigram_quantile = 0.8,
  min_freq = NULL,
  stem_collapse = TRUE,
  language = "English"
)
}
\arguments{
\item{x}{a vector or \code{data.frame} containing text}

\item{stop_words}{optional vector of strings, listing terms to be removed from the DTM prior to analysis. Defaults to synthesisr::get_stopwords()}

\item{bigram_check}{logical: should ngrams be searched for?}

\item{bigram_quantile}{what quantile of ngrams should be retained. Defaults to 0.8; i.e. the 80th percentile of bigram frequencies after removing all bigrams with frequencies <=2.}

\item{min_freq}{Numeric: the minimum number of times an ngram must occur to be included in the term list.}

\item{stem_collapse}{logical: should terms with identical lemmatized stems be merged?}

\item{language}{A string indicating which language should be used for stopwords.}
}
\value{
An object of class \code{simple_triplet_matrix}, listing the terms (columns) present in each row or string.
}
\description{
Takes bibliographic data and converts it to a DTM for passing to topic models.
}
\details{
This is primarily intended to be called internally by \code{screen_topics}, but is made available for users to generate their own topic models with the same properties as those in revtools.

This function uses some standard tools like stemming, converting words to lower case, and removal of numbers or punctuation. It also replaces stemmed words with the shortest version of all terms that share the same stem, which doesn't affect the calculations, but makes the resulting analyses easier to interpret. It doesn't use part-of-speech tagging.

Words that occur in 2 entries or fewer are always removed by \code{make_dtm}, so values of \code{min_freq} that result in a threshold below this will not affect the result. Arguments to \code{max_freq} are passed as is. Similarly words consisting of three letters or fewer are removed.

If \code{retain_empty_rows} is FALSE (the default) and the object returned is named \code{z}, then \code{as.numeric(z$dimnames$Docs)} provides an index to which entries have been retained from the input vector (\code{x}).
}
\examples{
my_df <-  data.frame(
  title = c(
    "EviAtlas: a tool for visualising evidence synthesis databases",

    "revtools: An R package to support article screening for evidence synthesis",

    "An automated approach to identifying search terms for
    systematic reviews using keyword co‐occurrence networks",

    "Reproducible, flexible and high‐throughput data extraction
    from primary literature: The metaDigitise r package"
  ),

  year = c("2019", "2019", "2019", "2019"),

  authors = c("Haddaway et al", "Westgate",
              "Grames et al", "Pick et al")
)


terms <- c("evidence synthesis", "package", "systematic")

create_dtm(as.character(my_df$title), bigram_check = TRUE, min_freq = 1)
}
