% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/data.R
\docType{data}
\name{databases}
\alias{databases}
\title{Information about databases recognized by synthesisr}
\format{A data frame with 18 obs of 3 variables

\describe{
 \item{database}{the name of the database}
 \item{signature}{a unique signature that appears in search results from this database}
 \item{import_function}{which import function to use if this database is detected}}}
\usage{
databases
}
\description{
A data frame containing databases, database signatures
used to recognize databases, and which function should
be called to import results from that database.
}
\keyword{datasets}
