% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/deduplication_functions.R
\name{deduplicate}
\alias{deduplicate}
\title{Remove duplicates from a bibliographic data set}
\usage{
deduplicate(x, matches)
}
\arguments{
\item{x}{A data.frame containing bibliographic data.}

\item{matches}{A character vector showing which entires in x are duplicates.}
}
\value{
Returns a data.frame of unique references.
}
\description{
Given a list of duplicate entries and a data set, this function extracts only unique references.
}
\examples{
my_df <-  data.frame(
  title = c(
    "EviAtlas: a tool for visualising evidence synthesis databases",

    "revtools: An R package to support article screening for evidence synthesis",

    "An automated approach to identifying search terms for
    systematic reviews using keyword co‐occurrence networks",

    "Reproducible, flexible and high‐throughput data extraction
    from primary literature: The metaDigitise r package",

    "eviatlas:tool for visualizing evidence synthesis databases.",

    "REVTOOLS a package to support article-screening for evidence synthsis"
  ),

  year = c("2019", "2019", "2019", "2019", NA, NA),

  authors = c("Haddaway et al", "Westgate",
              "Grames et al", "Pick et al", NA, NA)
)

dups <-
  find_duplicates(
    my_df,
    match_variable = "title",
    match_function = "stringdist",
    rm_punctuation = TRUE,
    to_lower = TRUE
  )

deduplicate(my_df, matches = dups)
}
