% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/import_functions.R
\name{detect_database}
\alias{detect_database}
\title{Detects from which database a search originated}
\usage{
detect_database(df)
}
\arguments{
\item{df}{a data frame of search results from a single source}
}
\value{
a character vector specifying the platform and/or database of origin, if known
}
\description{
Determines the database from which a search results file originated, if supported and exported per synthesisr instructions.
}
