% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/detect_format.R
\name{detect_delimiter}
\alias{detect_delimiter}
\title{Detect delimiter type in bibliographic files}
\usage{
detect_delimiter(x)
}
\arguments{
\item{x}{A character vector containing RIS-formatted bibliographic data.}
}
\value{
Returns the delimiter type used in a file.
}
\description{
The delimiter in ris files is often an endrow, a special character, or a space. This function detects which delimiter, if any, a file uses.
}
\examples{
file <- c(
  "TY  - JOUR",
  "AU  - Haddaway, Neal R.",
  "AU  - Feierman, Andrew",
  "AU  - Grainger, Matthew J.",
  "AU  - Gray, Charles T.",
  "AU  - Tanriver-Ayder, Ezgi",
  "AU  - Dhaubanjar, Sanita",
  "AU  - Westgate, Martin J.",
  "PY  - 2019",
  "DA  - 2019/06/04",
  "TI  - EviAtlas: a tool for visualising evidence synthesis databases",
  "JO  - Environmental Evidence",
  "SP  - 22",
  "VL  - 8",
  "IS  - 1",
  "SN  - 2047-2382",
  "UR  - https://doi.org/10.1186/s13750-019-0167-1",
  "DO  - 10.1186/s13750-019-0167-1",
  "ID  - Haddaway2019",
  "ER  - "
)

detect_delimiter(file)
}
