% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/detect_format.R
\name{detect_format}
\alias{detect_format}
\title{Detects if a file is bib-like or ris-like}
\usage{
detect_format(x)
}
\arguments{
\item{x}{A character vector containing bibliographic data.}
}
\value{
Returns the format of a file: either bib, ris, or unknown.
}
\description{
Because bibliographic data  can be stored in multiple file types, this function determines if the format of text is one of several likely probabilities: comma-separated; tab-separated; bibtex; and ris (medline or generic).
}
\examples{
metafor <- c("@article{viechtbauer2010,
  title={Conducting meta-analyses in R with the metafor package},
  author={Viechtbauer, Wolfgang},
  journal={Journal of Statistical Software},
  volume={36},
  number={3},
  pages={1--48},
  year={2010},
  publisher={UCLA Statistics}
}")

detect_format(metafor)
}
