% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/parsers.R
\name{detect_year}
\alias{detect_year}
\title{Fill in probable year values}
\usage{
detect_year(df)
}
\arguments{
\item{df}{a data.frame containing bibliographic data}
}
\value{
the input data.frame with years added for articles missing that information
}
\description{
Given a data.frame containing bibliographic data, this function attempts to detect the publication year based on date fields for articles missing a year.
}
\examples{
robvis <-
  c(
    "",
    "PMID- 32336025",
    "LR  - 20200426",
    "DP  - 2020 Apr 26",
    "TI  - Risk-Of-Bias VISualization (robvis): an R package and Shiny web app for",
    "      visualizing risk-of-bias assessments.",
    "LID - 10.1002/jrsm.1411 [doi]",
    "FAU - Higgins, Julian Pt",
    "LA  - eng",
    "PT  - Journal Article",
    "DEP - 20200426",
    "PL  - England",
    "TA  - Res Synth Methods",
    "JT  - Research Synthesis Methods",
    "OT  - Data visualization",
    "OT  - Evidence synthesis",
    "OT  - R",
    "OT  - Risk of bias",
    "EDAT- 2020/04/27 06:00",
    "MHDA- 2020/04/27 06:00",
    "CRDT- 2020/04/27 06:00",
    "PHST- 2020/02/27 00:00 [received]",
    "PHST- 2020/04/16 00:00 [revised]",
    "PHST- 2020/04/18 00:00 [accepted]",
    "PHST- 2020/04/27 06:00 [entrez]",
    "PHST- 2020/04/27 06:00 [pubmed]",
    "PHST- 2020/04/27 06:00 [medline]",
    "AID - 10.1002/jrsm.1411 [doi]",
    "SO  - Res Synth Methods. 2020 Apr 26. doi: 10.1002/jrsm.1411."
  )

tmp <- tempfile()

writeLines(robvis, tmp)

bib <- read_ref(tmp, return_df = TRUE, verbose = TRUE)
bib$year # data does not exist


df <- detect_year(bib)
df$year # likely publication years filled in based on dates found in other fields
}
