% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/find_duplicates.R
\name{find_duplicates}
\alias{find_duplicates}
\title{Detect duplicate values}
\usage{
find_duplicates(data, match_variable, group_variables, match_function,
  method, threshold, to_lower = FALSE, rm_punctuation = FALSE)
}
\arguments{
\item{data}{A data.frame that contains duplicate bibliographic entries.}

\item{match_variable}{A length-1 integer or string listing the column in which duplicates should be sought. Defaults to doi where available, followed by title. If neither are found the function will fail.}

\item{group_variables}{An optional vector listing the columns to use as grouping variables; that is, categories withing which duplicates should be sought (see 'note'). Optionally NULL to compare all entries against one another.}

\item{match_function}{The duplicate detection method to use; options are "stringdist" for similarity, "fuzzdist" for fuzzy matching, or "exact" for exact matches}

\item{method}{A string indicating the method to use for fuzzdist or stringdist.}

\item{threshold}{Numeric: the cutoff threshold for stringdist or fuzzdist.}

\item{to_lower}{Logical: Should all entries should be considered in lowercase when detecting duplicates?}

\item{rm_punctuation}{Logical: Should punctuation should be removed when detecting duplicates?}
}
\value{
Returns a vector of duplicate matches and methods used.
}
\description{
Identifies duplicate bibliographic entries using different duplicate detection methods.
}
\examples{
my_df <-  data.frame(
  title = c(
    "EviAtlas: a tool for visualising evidence synthesis databases",

    "revtools: An R package to support article screening for evidence synthesis",

    "An automated approach to identifying search terms for
    systematic reviews using keyword co‐occurrence networks",

    "Reproducible, flexible and high‐throughput data extraction
    from primary literature: The metaDigitise r package",

    "eviatlas:tool for visualizing evidence synthesis databases.",

    "REVTOOLS a package to support article-screening for evidence synthsis"
  ),

  year = c("2019", "2019", "2019", "2019", NA, NA),

  authors = c("Haddaway et al", "Westgate",
              "Grames et al", "Pick et al", NA, NA)
)

dups <-
  find_duplicates(
    my_df,
    match_variable = "title",
    match_function = "stringdist",
    rm_punctuation = TRUE,
    to_lower = TRUE
  )

deduplicate(my_df, matches = dups)
}
