% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/format_citation.R
\name{format_citation}
\alias{format_citation}
\alias{format_citation.list}
\alias{format_citation.bibliography}
\alias{format_citation.data.frame}
\title{Format a citation}
\usage{
format_citation(data, details = TRUE, abstract = FALSE,
  add_html = FALSE, line_breaks = FALSE, ...)

\method{format_citation}{list}(data, details = TRUE, abstract = FALSE,
  add_html = FALSE, line_breaks = FALSE, ...)

\method{format_citation}{bibliography}(data, details = TRUE,
  abstract = FALSE, add_html = FALSE, line_breaks = FALSE, ...)

\method{format_citation}{data.frame}(data, details = TRUE,
  abstract = FALSE, add_html = FALSE, line_breaks = FALSE, ...)
}
\arguments{
\item{data}{An object of class data.frame or bibliography.}

\item{details}{Logical: Should identifying information such as author names & journal titles be displayed? Defaults to TRUE.}

\item{abstract}{Logical: Should the abstract be shown (if available)? Defaults to FALSE.}

\item{add_html}{Logical: Should the journal title be italicized using html codes? Defaults to FALSE.}

\item{line_breaks}{Either logical, stating whether line breaks should be added, or numeric stating how many characters should separate consecutive line breaks. Defaults to FALSE.}

\item{...}{any other arguments.}
}
\value{
a string of length == length(x), containing formatted citations.
}
\description{
takes an object of class data.frame or bibliography and returns a formatted citation.
}
\section{Methods (by class)}{
\itemize{
\item \code{list}: Format citation from a list

\item \code{bibliography}: Format citation from a bibliography

\item \code{data.frame}: Format citation from a data.frame
}}

