% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/fuzz_functions.R
\name{fuzzdist}
\alias{fuzzdist}
\title{Detect duplicates with fuzzy matching}
\usage{
fuzzdist(a, b, method = c("fuzz_m_ratio", "fuzz_partial_ratio",
  "fuzz_token_sort_ratio", "fuzz_token_set_ratio"))
}
\arguments{
\item{a}{a character vector of items to match to b}

\item{b}{a character vector of items to match to a}

\item{method}{the method to use for fuzzy matching}
}
\value{
a vector of distances
}
\description{
This is a wrapper function for the different fuzzy matching methods
}
