% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/import_refs.R
\name{generate_bibliographic_names}
\alias{generate_bibliographic_names}
\title{Generate unique labels for entires}
\usage{
generate_bibliographic_names(x)
}
\arguments{
\item{x}{a list of bibliographic entires}
}
\value{
a character vector of unique names
}
\description{
Creates a unique label for each bibliographic entry using as much author and year data as possible.
}
