% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/dfm_functions.R
\name{get_tokens}
\alias{get_tokens}
\title{Retrieves tokens from a text}
\usage{
get_tokens(text, language)
}
\arguments{
\item{text}{A character vector containing text from which to extract tokens.}

\item{language}{A string indicating the language of the text.}
}
\value{
Returns a character vector of tokens from the text.
}
\description{
This function removes stopwords and extracts tokens from text.
}
\examples{
get_tokens("On the Origin of Species", language = "English")
}
