% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/database_import_functions.R
\name{import_MEDLINE}
\alias{import_MEDLINE}
\title{Import Web of Science MEDLINE}
\usage{
import_MEDLINE(df)
}
\arguments{
\item{df}{the data frame to import}
}
\value{
the data frame with synthesisr formatted columns
}
\description{
Imports files to synthesis from a data frame. Intended primarily as a function to be called by import_results()
}
