% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/database_import_functions.R
\name{import_OtherWoS}
\alias{import_OtherWoS}
\title{Import from miscellaneous Web of Science databases}
\usage{
import_OtherWoS(df)
}
\arguments{
\item{df}{the data frame to import}
}
\value{
the data frame with synthesisr formatted columns
}
\description{
Imports files to synthesis from a data frame. Intended primarily as a function to be called by import_results()
}
