% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/import_refs.R
\name{import_refs_internal}
\alias{import_refs_internal}
\title{Internal function called by import_refs for each file}
\usage{
import_refs_internal(filename, return_df = TRUE, verbose = TRUE)
}
\arguments{
\item{filename}{a path to a filename containing search results to import}

\item{return_df}{if TRUE, returns a dataframe; if FALSE, returns a list}

\item{verbose}{if TRUE, prints status updates}
}
\value{
a data frame of assembled search results
}
\description{
Workhorse function that imports bibliographic files; primarily intended to be called from import_refs.
}
