% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/import_functions.R
\name{import_results}
\alias{import_results}
\title{Import results of a systematic review}
\usage{
import_results(directory = NULL, filename = NULL, verbose = TRUE,
  retain_all_output = FALSE)
}
\arguments{
\item{directory}{a path to a directory containing search results to import}

\item{filename}{a path to a filename containing search results to import}

\item{verbose}{if TRUE, prints status updates}

\item{retain_all_output}{if TRUE, stores each search result as a list entry with original fields; if FALSE, merges all search output to a common 17 bibliographic fields}
}
\value{
a data frame of assembled search results
}
\description{
Given a file or directory, imports and assembles search results
}
