% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/deduplication_functions.R
\name{language_code}
\alias{language_code}
\title{Get short language codes}
\usage{
language_code(language)
}
\arguments{
\item{language}{a character vector containing the name of a language}
}
\value{
a character vector containing a two-letter language code
}
\description{
Returns the two-letter language code for specified language
}
