% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/deduplication_functions.R
\name{language_detect}
\alias{language_detect}
\title{Detect the language of a text}
\usage{
language_detect(text)
}
\arguments{
\item{text}{the text to which to assign a language}
}
\value{
a character vector of probable languages
}
\description{
Uses common stopwords to assign probable language(s) to a text
}
