% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/import_refs.R
\name{match_columns}
\alias{match_columns}
\title{Match imported data to reference codes}
\usage{
match_columns(df)
}
\arguments{
\item{df}{a data frame that contains bibliographic information}
}
\value{
a data frame rearranged and coded to match standard bibliographic fields, with other fields appended
}
\description{
Takes an imported dataframe, rearranges it to match lookup codes, and removes redundant columns
}
\examples{
files <-
  paste(system.file(package = "synthesisr"),
        "/extdata/zoorec.txt",
        sep = "")

df <- read.delim(files)
formatted <- match_columns(df)
colnames(formatted)
}
