% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/merge_columns.R
\name{merge_columns}
\alias{merge_columns}
\title{Bind two or more data frames with different columns}
\usage{
merge_columns(x, y)
}
\arguments{
\item{x}{either a data.frame or a list of data.frames}

\item{y}{a data.frame, optional if x is a list}
}
\value{
a data.frame
}
\description{
Takes two or more data.frames with different column names or
different column orders and binds them to a single data.frame.
}
