% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/generic_text_functions.R
\name{merge_stems}
\alias{merge_stems}
\title{Merge terms that have identical stems}
\usage{
merge_stems(dfm)
}
\arguments{
\item{dfm}{An object of class simple_triplet_matrix.}
}
\value{
The input dfm with duplicated terms merged into one column.
}
\description{
This function merges terms in document feature matrices if they have identical lemmatized stems.
}
