% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/deduplication_functions.R
\name{override_duplicates}
\alias{override_duplicates}
\title{Manually override duplicates}
\usage{
override_duplicates(matches, overrides)
}
\arguments{
\item{matches}{Numeric: a vector of group numbers for texts that indicates duplicates and unique values returned by the \code{\link{find_duplicates}} function.}

\item{overrides}{Numeric: a vector of group numbers that are not true duplicates.}
}
\value{
The input \code{matches} vector with unique group numbers for members of groups that the user overrides.
}
\description{
Re-assign group numbers to text that was classified as duplicated but is unique.
}
