% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/parsers.R
\name{parse_pubmed}
\alias{parse_pubmed}
\alias{parse_ris}
\alias{parse_bibtex}
\alias{parse_csv}
\alias{parse_tsv}
\title{Parse bibliographic text in a variety of formats}
\usage{
parse_pubmed(x)

parse_ris(x)

parse_bibtex(x)

parse_csv(x)

parse_tsv(x)
}
\arguments{
\item{x}{A character vector containing bibliographic information in ris format.}
}
\value{
Returns an object of class bibliography.
}
\description{
Text in standard formats - such as imported via readLines(), for example - can be parsed using a variety of standard formats. Use detect_format() to determine which is the most appropriate parser for your situation.
}
\section{Functions}{
\itemize{
\item \code{parse_pubmed}: Parse bibliographic text

\item \code{parse_bibtex}: Parse bibtex format

\item \code{parse_csv}: Parse comma-separated values

\item \code{parse_tsv}: Parse tab-separated values
}}

\examples{
file <- c(
  "TY  - JOUR",
  "AU  - Haddaway, Neal R.",
  "AU  - Feierman, Andrew",
  "AU  - Grainger, Matthew J.",
  "AU  - Gray, Charles T.",
  "AU  - Tanriver-Ayder, Ezgi",
  "AU  - Dhaubanjar, Sanita",
  "AU  - Westgate, Martin J.",
  "PY  - 2019",
  "DA  - 2019/06/04",
  "TI  - EviAtlas: a tool for visualising evidence synthesis databases",
  "JO  - Environmental Evidence",
  "SP  - 22",
  "VL  - 8",
  "IS  - 1",
  "SN  - 2047-2382",
  "UR  - https://doi.org/10.1186/s13750-019-0167-1",
  "DO  - 10.1186/s13750-019-0167-1",
  "ID  - Haddaway2019",
  "ER  - "
)

detect_format(file) # = "parse_ris"
df <- as.data.frame(parse_ris(file))
write_refs(df, format = "ris")
}
