% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/import_refs.R
\name{prep_ris}
\alias{prep_ris}
\title{Clean an ris file for import}
\usage{
prep_ris(z, delimiter)
}
\arguments{
\item{z}{A character vector that contains bibliographic information in ris format.}

\item{delimiter}{The type of delimiter separating entries.}
}
\value{
Returns a data.frame prepped for import with read_ris.
}
\description{
This function preps RIS files for import by cleaning common issues.
}
\examples{
file <- c(
  "TY  - JOUR",
  "AU  - Haddaway, Neal R.",
  "AU  - Feierman, Andrew",
  "AU  - Grainger, Matthew J.",
  "AU  - Gray, Charles T.",
  "AU  - Tanriver-Ayder, Ezgi",
  "AU  - Dhaubanjar, Sanita",
  "AU  - Westgate, Martin J.",
  "PY  - 2019",
  "DA  - 2019/06/04",
  "TI  - EviAtlas: a tool for visualising evidence synthesis databases",
  "JO  - Environmental Evidence",
  "SP  - 22",
  "VL  - 8",
  "IS  - 1",
  "SN  - 2047-2382",
  "UR  - https://doi.org/10.1186/s13750-019-0167-1",
  "DO  - 10.1186/s13750-019-0167-1",
  "ID  - Haddaway2019",
  "ER  - "
)

prep_ris(file, detect_delimiter(file))
}
