% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/import_refs.R
\name{read_bib}
\alias{read_bib}
\title{Read bib files}
\usage{
read_bib(x)
}
\arguments{
\item{x}{a character vector containing bibliographic information in bib format}
}
\value{
an object of class bibliography
}
\description{
Imports files using bib format.
}
\examples{
bibfile <- c(
  "@article{haddaway2019eviatlas,",
  "title={EviAtlas: a tool for visualising evidence synthesis databases},",
  "author={Haddaway, Neal R and Feierman, Andrew and Grainger, Matthew J
  and Gray, Charles T and Tanriver-Ayder, Ezgi and Dhaubanjar, Sanita
  and Westgate, Martin J},",
  "journal={Environmental Evidence},",
  "volume={8},",
  "number={1},",
  "pages={22},",
  "year={2019},",
  "publisher={BioMed Central}",
"}"
)

read_bib(bibfile)
}
