% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/import_functions.R
\name{read_files}
\alias{read_files}
\title{Reads in files to import}
\usage{
read_files(file)
}
\arguments{
\item{file}{a path to a file of a supported type}
}
\value{
a dataframe with the contents of the file
}
\description{
Given a path to a file, determines file type and reads it.
}
