% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/import_refs.R
\name{read_medline}
\alias{read_medline}
\title{Read medline-formatted ris files}
\usage{
read_medline(x)
}
\arguments{
\item{x}{a character vector containing bibliographic information in medline ris format}
}
\value{
an object of class bibliography
}
\description{
Imports files using medline ris format.
}
