% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/read_refs.R
\name{read_ref}
\alias{read_ref}
\title{Internal function called by read_refs for each file}
\usage{
read_ref(filename, return_df = TRUE, verbose = FALSE)
}
\arguments{
\item{filename}{A path to a filename containing search results to import.}

\item{return_df}{If TRUE, returns a data.frame; if FALSE, returns a list.}

\item{verbose}{If TRUE, prints status updates.}
}
\value{
Returns a data.frame or list of assembled search results.
}
\description{
This is the underlying workhorse function that imports bibliographic files; primarily intended to be called from read_refs.
}
\examples{
litsearchr <- c(
  "@article{grames2019,
  title={An automated approach to identifying search terms for
  systematic reviews using keyword co-occurrence networks},
  author={Grames, Eliza M and Stillman, Andrew N and Tingley, Morgan W and Elphick, Chris S},
  journal={Methods in Ecology and Evolution},
  volume={10},
  number={10},
  pages={1645--1654},
  year={2019},
  publisher={Wiley Online Library}
}"
)

tmp <- tempfile()

writeLines(litsearchr, tmp)

df <- read_refs(tmp, return_df = TRUE, verbose = TRUE)
}
