% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/import_refs.R
\name{read_ris}
\alias{read_ris}
\title{Read ris files}
\usage{
read_ris(x)
}
\arguments{
\item{x}{a character vector containing bibliographic information in ris format}
}
\value{
an object of class bibliography
}
\description{
Imports files using ris format.
}
