% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/read_refs.R
\name{remove_factors}
\alias{remove_factors}
\title{Remove factors from an object}
\usage{
remove_factors(z)
}
\arguments{
\item{z}{A data.frame}
}
\value{
Returns the input data.frame with all factors converted to character.
}
\description{
This function converts factors to characters to avoid errors with levels.
}
\examples{
remove_factors(list(as.factor(c("a", "b"))))
}
