% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/deduplication_functions.R
\name{remove_hyphens}
\alias{remove_hyphens}
\title{Remove hyphens from text}
\usage{
remove_hyphens(text)
}
\arguments{
\item{text}{the text from which to remove hyphens}
}
\value{
the input text with hyphens replaced with spaces
}
\description{
Replaces hyphens with a space
}
