% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/dfm_functions.R
\name{remove_numbers}
\alias{remove_numbers}
\title{Remove numbers from text}
\usage{
remove_numbers(text)
}
\arguments{
\item{text}{A character vector from which to remove numbers.}
}
\value{
Returns the input text with numbers removed.
}
\description{
Removes numbers from a text.
}
\examples{
remove_numbers("11s0y6nt4he35si6sr")
}
