% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/dfm_functions.R
\name{remove_stopwords}
\alias{remove_stopwords}
\title{Remove stopwords from text}
\usage{
remove_stopwords(text, language)
}
\arguments{
\item{text}{the text from which to remove stopwords}

\item{language}{the language of the text}
}
\value{
the text with stopwords removed
}
\description{
Removes stopwords from text in user-specified language
}
\examples{
remove_stopwords("On the Origin of Species", language="English")
}
