% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/reporting_functions.R
\name{report_accumulation}
\alias{report_accumulation}
\title{Report accumulation of new results per database}
\usage{
report_accumulation(df, database_column, database_order = NULL)
}
\arguments{
\item{df}{the data frame of search results}

\item{database_column}{the name or index of the column where database information is contained}

\item{database_order}{a character vector of database names that appear in the database column, in the order that they should be added to the accumulation curve}
}
\value{
a data frame with an accumulation of unique hits per unit of searching effort
}
\description{
Generates an accumulation curve of new unique hits per database
}
