% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/reporting_functions.R
\name{report_results}
\alias{report_results}
\title{Report search results per database}
\usage{
report_results(df, database_column)
}
\arguments{
\item{df}{the data frame of search results}

\item{database_column}{the name or index of the column where database information is contained}
}
\value{
a data frame with a count of results per database
}
\description{
Reports how many hits were imported per database
}
