% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/deduplication_functions.R
\name{review_duplicates}
\alias{review_duplicates}
\title{Manually review potential duplicates}
\usage{
review_duplicates(text, matches)
}
\arguments{
\item{text}{A character vector of the text that was used to identify potential duplicates.}

\item{matches}{Numeric: a vector of group numbers for texts that indicates duplicates and unique values returned by the \code{\link{find_duplicates}} function.}
}
\value{
A \code{data.frame} of potential duplicates grouped together.
}
\description{
Allows users to manually review articles classified as duplicates.
}
