% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/import_refs.R
\name{rollingsum}
\alias{rollingsum}
\title{Computes a rolling sum}
\usage{
rollingsum(a, n = 2L)
}
\arguments{
\item{a}{A string or character vector.}

\item{n}{Numeric: the window of previous detections to consider.}
}
\value{
Returns a vector of numbers passed to detect_delimiter.
}
\description{
This function computes the rolling sum of detections; intended for use in detect_delimiter.
}
\examples{
rollingsum(c(1,1,2,3,5))
}
