% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/import_functions.R
\name{standardize_df}
\alias{standardize_df}
\title{Convert dataframes to standard format}
\usage{
standardize_df(df)
}
\arguments{
\item{df}{a data frame}
}
\value{
a data frame with standardized columns
}
\description{
Given a data frame, checks to see that it contains standard synthesisr fields and fills empty fields.
}
