% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/string_.R, R/string_functions.R
\name{string_}
\alias{string_}
\alias{string_osa}
\alias{string_lv}
\alias{string_dl}
\alias{string_hamming}
\alias{string_lcs}
\alias{string_qgram}
\alias{string_cosine}
\alias{string_jaccard}
\alias{string_jw}
\alias{string_soundex}
\title{Calculate similarity between two strings}
\usage{
string_osa(a, b)

string_lv(a, b)

string_dl(a, b)

string_hamming(a, b)

string_lcs(a, b)

string_qgram(a, b)

string_cosine(a, b)

string_jaccard(a, b)

string_jw(a, b)

string_soundex(a, b)
}
\arguments{
\item{a}{A character vector of items to match to b.}

\item{b}{A character vector of items to match to a.}
}
\value{
Returns a score of same length as b, giving the dissimilarity between a and b.
}
\description{
These functions each access a specific \code{"methods"} argument provided by \code{stringdist}, and are provided for convenient calling by \code{\link{find_duplicates}}. They do not include any new functionality beyond that given by \code{stringdist}, which you should use for your own analyses.
}
