% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/synthesisr-package.R
\docType{package}
\name{synthesisr-package}
\alias{synthesisr}
\alias{synthesisr-package}
\title{synthesisr: Import, assemble, and deduplicate bibiliographic datasets}
\description{
Systematic review searches include multiple databases
that export results in a variety of formats with overlap in
coverage between databases. To streamline the process of importing,
assembling, and deduplicating results, \code{synthesisr} recognizes
bibliographic files exported from databases commonly used for
systematic reviews and merges results into a standardized format.
}
\section{Import & Export}{

The key task performed by \code{synthesisr} is flexible import and
presentation of bibliographic data. This is typically achieved by
\code{read_refs()}, which can import multiple files at once and link them together
into a single \code{tibble}. Conversely, export is via \code{write_refs()}. Users
that require more detailed control can use the following functions:

\itemize{
\item \link{read_refs} Read bibliographic data
\item \link{write_refs} Write bibliographic data
\item \link{detect_} Detect file attributes
\item \link{parse_} Parse a vector containing bibliographic data
\item \link{clean_} Cleaning functions for author and column names
\item \link{code_lookup} A dataset of potential ris tags
}
}

\section{Formatting}{

\itemize{
\item \linkS4class{bibliography} Methods for class \code{bibliography}
\item \link{format_citation} Return a clean citation from a \code{bibliography} or \code{data.frame}
\item \link{add_line_breaks} Set a maximum character width for strings
}
}

\section{Deduplication}{

When importing from multiple databases, it is likely that there will be
duplicates in the resulting dataset. The easiest way to deal with this
problem in \code{synthesisr} is using the \code{deduplicate()} function; but this can
be risky, particularly if there are no DOIs in the dataset. To get finer
control of the deduplication process, consider using the sub-functions:

\itemize{
\item \link{deduplicate} Semi-automated duplicate removal
\item \link{find_duplicates} Locate potentially duplicated references
\item \link{extract_unique_references} Return a data.frame with only 'unique' references
\item \link{review_duplicates} Manually review potential duplicates
\item \link{override_duplicates} Manually override identified duplicates
\item \link{fuzz_} Fuzzy string matching c/o \code{fuzzywuzzy}
\item \link{string_} Fuzzy string matching c/o \code{stringdist}
}
}

\seealso{
Useful links:
\itemize{
  \item \url{https://martinwestgate.com/synthesisr/}
  \item Report bugs at \url{https://github.com/mjwestgate/synthesisr/issues}
}

}
\author{
\strong{Maintainer}: Martin Westgate \email{martinjwestgate@gmail.com} (\href{https://orcid.org/0000-0003-0854-2034}{ORCID})

Authors:
\itemize{
  \item Eliza Grames \email{eliza.grames@uconn.edu} (\href{https://orcid.org/0000-0003-1743-6815}{ORCID})
}

}
