% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/write_bibliography.R
\name{write_bib}
\alias{write_bib}
\title{Write a .bib file}
\usage{
write_bib(x)
}
\arguments{
\item{x}{An object of class bibliography.}
}
\value{
Returns a character vector containing references in .bib format.
}
\description{
This is a subfunction of write_bibliography to convert an object of class bibliography to .bib text format.
}
\examples{
bibfile <- c(
  "@article{haddaway2019eviatlas,",
  "title={EviAtlas: a tool for visualising evidence synthesis databases},",
  "author={Haddaway, Neal R and Feierman, Andrew and Grainger, Matthew J
  and Gray, Charles T and Tanriver-Ayder, Ezgi and Dhaubanjar, Sanita
  and Westgate, Martin J},",
  "journal={Environmental Evidence},",
  "volume={8},",
  "number={1},",
  "pages={22},",
  "year={2019},",
  "publisher={BioMed Central}",
"}"
)

bib <- read_bib(bibfile)

write_bibliography(bib, format = "bib")
}
