% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/write_bibliography.R
\name{write_bibliography}
\alias{write_bibliography}
\title{Export data to a bibliographic format}
\usage{
write_bibliography(x, format = "ris", write_file = FALSE,
  filename = NULL)
}
\arguments{
\item{x}{a data.frame containing bibliographic information or an object of class bibliography}

\item{format}{what format to export data as; options are "ris" or "bib"}

\item{write_file}{if TRUE, saves the result to a text file in the working directory}

\item{filename}{if write_file is TRUE, the name of the file to be written}
}
\value{
a character vector containing bibliographic information in the specified format if write_file is FALSE
}
\description{
Exports data.frames containing bibliographic information to either a .ris or .bib file.
}
