% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/write_refs.R
\name{write_refs}
\alias{write_refs}
\alias{write_bib}
\alias{write_ris}
\title{Export data to a bibliographic format}
\usage{
write_refs(x, file, format = "ris", tag_naming = "synthesisr", write = TRUE)

write_bib(x)

write_ris(x, tag_naming = "synthesisr")
}
\arguments{
\item{x}{Either a \code{tibble} containing bibliographic information, or an
object of class \code{bibliography.}}

\item{file}{filename to save to.}

\item{format}{What format should the data be exported as? Options are \code{"ris"} or
\code{"bib"}.}

\item{tag_naming}{what naming convention should be used to write RIS files?
See details for options.}

\item{write}{Logical should a file should be written? If \code{FALSE} returns a
\code{list}.}
}
\value{
This function is typically called for it's side effect of writing a
file in the specified location and format. If \code{write} is FALSE, returns
a character vector containing bibliographic information in the specified
format.
}
\description{
This function exports data.frames containing bibliographic
information to either a \code{.ris} or \code{.bib} file.
}
\examples{
eviatlas <- c(
  "TY  - JOUR",
  "AU  - Haddaway, Neal R.",
  "AU  - Feierman, Andrew",
  "AU  - Grainger, Matthew J.",
  "AU  - Gray, Charles T.",
  "AU  - Tanriver-Ayder, Ezgi",
  "AU  - Dhaubanjar, Sanita",
  "AU  - Westgate, Martin J.",
  "PY  - 2019",
  "DA  - 2019/06/04",
  "TI  - EviAtlas: a tool for visualising evidence synthesis databases",
  "JO  - Environmental Evidence",
  "SP  - 22",
  "VL  - 8",
  "IS  - 1",
  "SN  - 2047-2382",
  "UR  - https://doi.org/10.1186/s13750-019-0167-1",
  "DO  - 10.1186/s13750-019-0167-1",
  "ID  - Haddaway2019",
  "ER  - "
)

detect_parser(eviatlas) # = "parse_ris"
df <- parse_ris(eviatlas) |>
  as_tibble()
ris_out <- write_refs(df, format = "ris", write = FALSE)
}
