% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/write_bibliography.R
\name{write_ris}
\alias{write_ris}
\title{Write a .ris file}
\usage{
write_ris(x)
}
\arguments{
\item{x}{an object of class bibliography}
}
\value{
a character vector containing references in .ris format
}
\description{
a subfunction of write_bibliography to convert a bibliography to .ris format
}
