% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/add_line_breaks.R
\name{add_line_breaks}
\alias{add_line_breaks}
\title{Add line breaks to one or more strings}
\usage{
add_line_breaks(x, n = 50, html = FALSE)
}
\arguments{
\item{x}{Either a string or a vector; if the vector is not of class character
if will be coerced to one using \code{as.character()}.}

\item{n}{Numeric: The desired number of characters that should separate
consecutive line breaks.}

\item{html}{Logical: Should the line breaks be specified in html?}
}
\value{
Returns the input vector unaltered except for the addition of line
breaks.
}
\description{
This function takes a vector of strings and adds line breaks
every n characters. Primarily built to be called internally by
\code{format_citation()}, this function has been made available as it can be
useful in other contexts.
}
\details{
Line breaks are only added between words, so the value of n is
actually a threshold value rather than being matched exactly.
}
\examples{
add_line_breaks(c("On the Origin of Species"), n = 10)
}
