% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/clean_df.R
\name{clean_df}
\alias{clean_df}
\title{Cleans data.frames into synthesisr format}
\usage{
clean_df(data)
}
\arguments{
\item{data}{A data.frame with bibliographic information.}
}
\value{
Returns the input data.frame cleaned and standardized.
}
\description{
Cleans column names and labels in data.frames into a format suitable for synthesisr applications.
}
\examples{
my_df <-  data.frame(
  X..title. = c(
    "EviAtlas: a tool for visualising evidence synthesis databases",

    "revtools: An R package to support article screening for evidence synthesis",

    "An automated approach to identifying search terms for
      systematic reviews using keyword co‐occurrence networks",

    "Reproducible, flexible and high‐throughput data extraction
    from primary literature: The metaDigitise r package"),

  YEAR = c("2019", "2019", "2019", "2019"),

  authors = c("Haddaway et al", "Westgate",
              "Grames et al", "Pick et al"))

clean_df(my_df)
}
\seealso{
[clean_authors()] and [clean_names()] for underlying functions.
}
