% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/clean_df.R
\name{clean_names}
\alias{clean_names}
\title{Clean common issues with column names}
\usage{
clean_names(x)
}
\arguments{
\item{x}{A string or vector of column names.}
}
\value{
Returns the input vector with common issues resolved.
}
\description{
Cleans column names from imports such as extra punctuation.
}
\examples{
clean_names(c(".title...", "X..YEAR", "authors.."))
}
