% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/dfm_functions.R
\name{create_dfm}
\alias{create_dfm}
\title{Create a document-feature matrix}
\usage{
create_dfm(elements, features, closure = c("left", "right", "full",
  "none", to_lower = TRUE))
}
\arguments{
\item{elements}{a character vector of document information (e.g. document titles or abstracts)}

\item{features}{a character vector of terms to use as document features}

\item{closure}{restrictions on how keywords are detected; left requires terms to start with a keyword (e.g "burn" matches "burning"), right requires terms to end with a keyword (e.g. "burn" matches "postburn" but not "postburning"), full requires exact matches (e.g. "burn" only matches "burn"), and none allows keywords to be embedded within terms.}

\item{ignore_case}{if case be ignored when detecting features within documents}
}
\value{
a matrix with documents as rows and terms as columns
}
\description{
Given a character vector of document information, creates a document-feature matrix.
}
