% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/clean_df.R
\name{create_index}
\alias{create_index}
\title{Create a string to sort in correct order}
\usage{
create_index(string, n, sep = "_")
}
\arguments{
\item{string}{a character vector}

\item{n}{the number of unique index values to create}
}
\value{
a character vector
}
\description{
Creates a string of named length in format "string_number" to index in correct order.
}
