% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/deduplication_functions.R
\name{deduplicate}
\alias{deduplicate}
\title{Remove duplicate entries from a data frame}
\usage{
deduplicate(df, field, method = c("quick", "similarity", "fuzzy"),
  language = "English", cutoff_distance = 2)
}
\arguments{
\item{df}{the data frame to deduplicate}

\item{field}{the name or index of the column to check for duplicate values}

\item{method}{the manner of duplicate detection; quick removes exact text duplicates, similarity removes duplicates below a similarity threshold, and fuzzy uses fuzzdist matching}

\item{language}{the language to use if method is set to similarity}

\item{cutoff_distance}{the threshold below which articles are marked as duplicates by the similarity method}
}
\value{
a deduplicated data frame
}
\description{
Given a data frame and a field to check for duplicates, flags and removes duplicate entries with three optional methods.
}
