% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/extract_unique_references.R
\name{deduplicate}
\alias{deduplicate}
\title{Remove duplicates from a bibliiographic data set}
\usage{
deduplicate(x, matches)
}
\arguments{
\item{x}{a data.frame in the format returned by make_reference_dataframe()}

\item{matches}{a character vector showing which values are duplicates}
}
\value{
a data.frame of unique references
}
\description{
Given a list of duplicate entries and a data set, extracts only unique references
}
