% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/detect_format.R
\name{detect_delimiter}
\alias{detect_delimiter}
\title{Detect delimiter type in bibliographic files}
\usage{
detect_delimiter(x)
}
\arguments{
\item{x}{A character vector containing RIS-formatted bibliographic data.}
}
\value{
Returns the delimiter type used in a file.
}
\description{
The delimiter in ris files is often an endrow, a special character, or a space. This function detects which delimiter, if any, a file uses.
}
\examples{
revtools <- c("",
              "PMID- 31355546",
              "OWN - NLM",
              "STAT- In-Process",
              "LR  - 20200226",
              "IS  - 1759-2887 (Electronic)",
              "IS  - 1759-2879 (Linking)",
              "VI  - 10",
              "IP  - 4",
              "DP  - 2019 Dec",
              "TI  - revtools: An R package to support article
                     screening for evidence synthesis.",
              "PG  - 606-614",
              "LID - 10.1002/jrsm.1374 [doi]",
              "AU  - Westgate MJ",
              "LA  - eng",
              "PT  - Journal Article",
              "DEP - 20191018",
              "TA  - Res Synth Methods",
              "JT  - Research Synthesis Methods",
              "JID - 101543738",
              "SB  - IM",
              "OTO - NOTNLM",
              "OT  - data visualization",
              "OT  - meta-analysis",
              "OT  - natural language processing",
              "OT  - systematic review",
              "OT  - topic models",
              "EDAT- 2019/07/30 06:00",
              "MHDA- 2019/07/30 06:00",
              "CRDT- 2019/07/30 06:00",
              "PHST- 2019/02/25 00:00 [received]",
              "PHST- 2019/06/12 00:00 [revised]",
              "PHST- 2019/07/23 00:00 [accepted]",
              "AID - 10.1002/jrsm.1374 [doi]",
              "PST - ppublish",
              "      18."
)

detect_delimiter(revtools)
}
