% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/import_refs.R
\name{detect_delimiter}
\alias{detect_delimiter}
\title{Detect delimiter type in bibliographic files}
\usage{
detect_delimiter(x)
}
\arguments{
\item{x}{a bibliographic file}
}
\value{
the delimiter type
}
\description{
Detects if the delimiter in a file is endrow, character, or space
}
\examples{
inst/examples/detect_delimiter.R
}
