% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/import_refs.R
\name{detect_filetype}
\alias{detect_filetype}
\title{Detects file types}
\usage{
detect_filetype(filename)
}
\arguments{
\item{filename}{a path to a file}
}
\value{
a character vector with the likely file type based on the file extension
}
\description{
Given a file, determines the file extension
}
