% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/import_refs.R
\name{detect_filetype}
\alias{detect_filetype}
\title{Detects file types}
\usage{
detect_filetype(filename)
}
\arguments{
\item{filename}{A path to a file.}
}
\value{
Returns a character vector with the likely file type based on the file extension and content.
}
\description{
Because different file types require different import strategies, this function detects the input file type.
}
\examples{
files <-
  paste(system.file(package = "synthesisr"),
        "/extdata/zoorec.txt",
        sep = "")

detect_filetype(files)
}
