% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/read_refs.R
\name{detect_format}
\alias{detect_format}
\title{Detects if a file is bib-like or ris-like}
\usage{
detect_format(x)
}
\arguments{
\item{x}{A character vector containing bibliographic data.}
}
\value{
Returns the format of a file: either bib, ris, or unknown.
}
\description{
Because bibliographic data  can be stored in multiple file types, this function determines if the format of text is more bib-like or ris-like.
}
\examples{
file <- c(
  "TY  - JOUR",
  "AU  - Haddaway, Neal R.",
  "AU  - Feierman, Andrew",
  "AU  - Grainger, Matthew J.",
  "AU  - Gray, Charles T.",
  "AU  - Tanriver-Ayder, Ezgi",
  "AU  - Dhaubanjar, Sanita",
  "AU  - Westgate, Martin J.",
  "PY  - 2019",
  "DA  - 2019/06/04",
  "TI  - EviAtlas: a tool for visualising evidence synthesis databases",
  "JO  - Environmental Evidence",
  "SP  - 22",
  "VL  - 8",
  "IS  - 1",
  "SN  - 2047-2382",
  "UR  - https://doi.org/10.1186/s13750-019-0167-1",
  "DO  - 10.1186/s13750-019-0167-1",
  "ID  - Haddaway2019",
  "ER  - "
)
detect_format(file)
}
