% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/parsers.R
\name{detect_year}
\alias{detect_year}
\title{Fill in probable year values}
\usage{
detect_year(df)
}
\arguments{
\item{df}{a data.frame containing bibliographic data}
}
\value{
the input data.frame with years added for articles missing that information
}
\description{
Given a data.frame containing bibliographic data, this function attempts to detect the publication year based on date fields for articles missing a year.
}
\examples{
res_synth_methods <-
  c(
    "",
    "PMID- 32336025",
    "OWN - NLM",
    "STAT- Publisher",
    "LR  - 20200426",
    "IS  - 1759-2887 (Electronic)",
    "IS  - 1759-2879 (Linking)",
    "DP  - 2020 Apr 26",
    "TI  - Risk-Of-Bias VISualization (robvis): an R package and Shiny web app for",
    "      visualizing risk-of-bias assessments.",
    "LID - 10.1002/jrsm.1411 [doi]",
    "CI  - This article is protected by copyright. All rights reserved.",
    "FAU - McGuinness, Luke A",
    "AU  - McGuinness LA",
    "AUID- ORCID: https://orcid.org/0000-0001-8730-9761",
    "AD  - MRC Integrative Epidemiology Unit at the University of Bristol, Bristol, UK.",
    "AD  - Population Health Sciences, Bristol Medical School, University of Bristol,",
    "      Bristol, UK.",
    "FAU - Higgins, Julian Pt",
    "AU  - Higgins JP",
    "AD  - MRC Integrative Epidemiology Unit at the University of Bristol, Bristol, UK.",
    "AD  - Population Health Sciences, Bristol Medical School, University of Bristol,",
    "      Bristol, UK.",
    "LA  - eng",
    "PT  - Journal Article",
    "DEP - 20200426",
    "PL  - England",
    "TA  - Res Synth Methods",
    "JT  - Research synthesis methods",
    "JID - 101543738",
    "SB  - IM",
    "OTO - NOTNLM",
    "OT  - Data visualization",
    "OT  - Evidence synthesis",
    "OT  - R",
    "OT  - Risk of bias",
    "EDAT- 2020/04/27 06:00",
    "MHDA- 2020/04/27 06:00",
    "CRDT- 2020/04/27 06:00",
    "PHST- 2020/02/27 00:00 [received]",
    "PHST- 2020/04/16 00:00 [revised]",
    "PHST- 2020/04/18 00:00 [accepted]",
    "PHST- 2020/04/27 06:00 [entrez]",
    "PHST- 2020/04/27 06:00 [pubmed]",
    "PHST- 2020/04/27 06:00 [medline]",
    "AID - 10.1002/jrsm.1411 [doi]",
    "PST - aheadofprint",
    "SO  - Res Synth Methods. 2020 Apr 26. doi: 10.1002/jrsm.1411.",
    "",
    "PMID- 31355546",
    "OWN - NLM",
    "STAT- In-Process",
    "LR  - 20200226",
    "IS  - 1759-2887 (Electronic)",
    "IS  - 1759-2879 (Linking)",
    "VI  - 10",
    "IP  - 4",
    "DP  - 2019 Dec",
    "TI  - revtools: An R package to support article screening for evidence synthesis.",
    "PG  - 606-614",
    "LID - 10.1002/jrsm.1374 [doi]",
    "FAU - Westgate, Martin J",
    "AU  - Westgate MJ",
    "AUID- ORCID: https://orcid.org/0000-0003-0854-2034",
    "AD  - Fenner School of Environment & Society, The Australian National University,",
    "      Acton, ACT, Australia.",
    "LA  - eng",
    "PT  - Journal Article",
    "DEP - 20191018",
    "PL  - England",
    "TA  - Res Synth Methods",
    "JT  - Research synthesis methods",
    "JID - 101543738",
    "SB  - IM",
    "OTO - NOTNLM",
    "OT  - data visualization",
    "OT  - meta-analysis",
    "OT  - natural language processing",
    "OT  - systematic review",
    "OT  - topic models",
    "EDAT- 2019/07/30 06:00",
    "MHDA- 2019/07/30 06:00",
    "CRDT- 2019/07/30 06:00",
    "PHST- 2019/02/25 00:00 [received]",
    "PHST- 2019/06/12 00:00 [revised]",
    "PHST- 2019/07/23 00:00 [accepted]",
    "PHST- 2019/07/30 06:00 [pubmed]",
    "PHST- 2019/07/30 06:00 [medline]",
    "PHST- 2019/07/30 06:00 [entrez]",
    "AID - 10.1002/jrsm.1374 [doi]",
    "PST - ppublish",
    "SO  - Res Synth Methods. 2019 Dec;10(4):606-614. doi: 10.1002/jrsm.1374. Epub 2019 Oct ",
    "      18."
  )

tmp <- tempfile(pattern = "res_synth_methods")

writeLines(res_synth_methods, tmp)

bib <- read_ref(tmp, return_df = TRUE, verbose = TRUE)
bib$year # data does not exist

df <- detect_year(bib)
df$year # likely publication years filled in based on dates found in other fields
}
