% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/format_citation.R
\name{format_citation}
\alias{format_citation}
\alias{format_citation.list}
\alias{format_citation.bibliography}
\alias{format_citation.data.frame}
\title{Format a citation}
\usage{
format_citation(data, details = TRUE, abstract = FALSE,
  add_html = FALSE, line_breaks = FALSE, ...)

\method{format_citation}{list}(data, details = TRUE, abstract = FALSE,
  add_html = FALSE, line_breaks = FALSE, ...)

\method{format_citation}{bibliography}(data, details = TRUE,
  abstract = FALSE, add_html = FALSE, line_breaks = FALSE, ...)

\method{format_citation}{data.frame}(data, details = TRUE,
  abstract = FALSE, add_html = FALSE, line_breaks = FALSE, ...)
}
\arguments{
\item{data}{An object of class data.frame or bibliography.}

\item{details}{Logical: Should identifying information such as author names & journal titles be displayed? Defaults to TRUE.}

\item{abstract}{Logical: Should the abstract be shown (if available)? Defaults to FALSE.}

\item{add_html}{Logical: Should the journal title be italicized using html codes? Defaults to FALSE.}

\item{line_breaks}{Either logical, stating whether line breaks should be added, or numeric stating how many characters should separate consecutive line breaks. Defaults to FALSE.}

\item{...}{any other arguments.}
}
\value{
Returns a string of length equal to length(x) that contains formatted citations.
}
\description{
This function takes an object of class data.frame or bibliography and returns a formatted citation.
}
\section{Methods (by class)}{
\itemize{
\item \code{list}: Format citation from a list

\item \code{bibliography}: Format citation from a bibliography

\item \code{data.frame}: Format citation from a data.frame
}}

\examples{
file <- c(
  "TY  - JOUR",
  "AU  - Haddaway, Neal R.",
  "AU  - Feierman, Andrew",
  "AU  - Grainger, Matthew J.",
  "AU  - Gray, Charles T.",
  "AU  - Tanriver-Ayder, Ezgi",
  "AU  - Dhaubanjar, Sanita",
  "AU  - Westgate, Martin J.",
  "PY  - 2019",
  "DA  - 2019/06/04",
  "TI  - EviAtlas: a tool for visualising evidence synthesis databases",
  "JO  - Environmental Evidence",
  "SP  - 22",
  "VL  - 8",
  "IS  - 1",
  "SN  - 2047-2382",
  "UR  - https://doi.org/10.1186/s13750-019-0167-1",
  "DO  - 10.1186/s13750-019-0167-1",
  "ID  - Haddaway2019",
  "ER  - "
)

delim <- detect_delimiter(file)
citation <- read_ris(prep_ris(file, delim))
format_citation(citation)
}
