% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/fuzz_functions.R
\name{fuzzdist}
\alias{fuzzdist}
\alias{fuzz_m_ratio}
\alias{fuzz_partial_ratio}
\alias{fuzz_token_sort_ratio}
\alias{fuzz_token_set_ratio}
\title{Detect duplicates with fuzzy matching}
\usage{
fuzzdist(a, b, method = c("fuzz_m_ratio", "fuzz_partial_ratio",
  "fuzz_token_sort_ratio", "fuzz_token_set_ratio"))

fuzz_m_ratio(a, b)

fuzz_partial_ratio(a, b)

fuzz_token_sort_ratio(a, b)

fuzz_token_set_ratio(a, b)
}
\arguments{
\item{a}{a character vector of items to match to b}

\item{b}{a character vector of items to match to a}

\item{method}{the method to use for fuzzy matching}
}
\value{
a vector of distances
}
\description{
This is a wrapper function for the different fuzzy matching methods
}
\section{Functions}{
\itemize{
\item \code{fuzz_m_ratio}: Method simple ratio

\item \code{fuzz_partial_ratio}: Method partial ratio

\item \code{fuzz_token_sort_ratio}: Method token sort ratio

\item \code{fuzz_token_set_ratio}: Method token set ratio
}}

