% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/dfm_functions.R
\name{get_stopwords}
\alias{get_stopwords}
\title{Retrieve stopwords for a given language}
\usage{
get_stopwords(language)
}
\arguments{
\item{language}{a character vector containing the name of the language for which to retrieve stopwords}
}
\value{
a character vector of stopwords
}
\description{
Returns a character vector of stopwords to use for a specified language
}
