% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/deduplication_functions.R
\name{get_tokens}
\alias{get_tokens}
\title{Retrieves tokens from a text}
\usage{
get_tokens(text, language)
}
\arguments{
\item{text}{the text from which to remove stopwords}

\item{language}{the language of the text}
}
\value{
a character vector of tokens from the text
}
\description{
Removes stopwords and extracts tokens from text
}
