% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/import_refs.R
\name{import_refs}
\alias{import_refs}
\title{Import bibliographic search results}
\usage{
import_refs(filename, return_df = TRUE, verbose = TRUE)
}
\arguments{
\item{filename}{A path to a filename or vector of filenames containing search results to import.}

\item{return_df}{If TRUE, returns a data.frame; if FALSE, returns a list.}

\item{verbose}{If TRUE, prints status updates.}
}
\value{
Returns a data.frame or list of assembled search results.
}
\description{
Imports common bibliographic reference formats (i.e. .bib, .ris, or .txt).
}
\examples{
files <-
  paste(system.file(package = "synthesisr"),
        "/extdata/zoorec.txt",
        sep = "")
results <- import_refs(files)
}
