% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/import_functions.R
\name{import_results}
\alias{import_results}
\title{Import results of a systematic review}
\usage{
import_results(directory = NULL, filename = NULL,
  save_dataset = FALSE, save_directory = "./", verbose = TRUE)
}
\arguments{
\item{directory}{a path to a directory containing search results to import}

\item{filename}{a path to a filename containing search results to import}

\item{save_dataset}{if TRUE, saves the full search results to a .csv}

\item{save_directory}{the path to a directory where search results will be saved if save_dataset is set to TRUE}

\item{verbose}{if TRUE, prints status updates}
}
\value{
a data frame of assembled search results
}
\description{
Given a file or directory, imports and assembles search results
}
