% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/read_refs.R
\name{match_columns}
\alias{match_columns}
\title{Matches imported data to reference codes}
\usage{
match_columns(df)
}
\arguments{
\item{df}{A data.frame that contains bibliographic information.}
}
\value{
Returns a data.frame rearranged and coded to match standard bibliographic fields, with unrecognized fields appended.
}
\description{
Takes an imported data.frame and rearranges it to match lookup codes.
}
\examples{
files <-
  paste(system.file(package = "synthesisr"),
        "/extdata/zoorec.txt",
        sep = "")

df <- read.delim(files)
formatted <- match_columns(df)
colnames(formatted)
}
