% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/import_refs.R
\name{prep_ris}
\alias{prep_ris}
\title{Clean an RIS file for import}
\usage{
prep_ris(z, delimiter)
}
\arguments{
\item{z}{an character vector that contains RIS bibliographic information}

\item{delimiter}{the type of delimiter separating entries}
}
\value{
a data.frame intended for import with read_ris
}
\description{
Preps RIS files by cleaning common issues and converting to a common format
}
