% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/parsers.R
\name{prep_ris}
\alias{prep_ris}
\title{Clean an RIS file for import}
\usage{
prep_ris(z, delimiter)
}
\arguments{
\item{z}{A character vector that contains RIS bibliographic information.}

\item{delimiter}{A string indicating the type of delimiter separating entries.}
}
\value{
Returns a data.frame intended for import with parse_ris.
}
\description{
This function preps RIS files by cleaning common issues and converting to a common format.
}
