% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/import_refs.R
\name{read_bib}
\alias{read_bib}
\title{Read bib files}
\usage{
read_bib(x)
}
\arguments{
\item{x}{A character vector containing bibliographic information in bib format.}
}
\value{
Returns an object of class bibliography.
}
\description{
This function reads in bibliographic data stored in .bib format.
}
\examples{
bibfile <- c(
  "@article{haddaway2019eviatlas,",
  "title={EviAtlas: a tool for visualising evidence synthesis databases},",
  "author={Haddaway, Neal R and Feierman, Andrew and Grainger, Matthew J
  and Gray, Charles T and Tanriver-Ayder, Ezgi and Dhaubanjar, Sanita
  and Westgate, Martin J},",
  "journal={Environmental Evidence},",
  "volume={8},",
  "number={1},",
  "pages={22},",
  "year={2019},",
  "publisher={BioMed Central}",
"}"
)

bib <- read_bib(bibfile)

write_bibliography(bib, format = "bib")
}
