% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/import_refs.R
\name{read_bib}
\alias{read_bib}
\title{Read bib files}
\usage{
read_bib(x)
}
\arguments{
\item{x}{a character vector containing bibliographic information in bib format}
}
\value{
an object of class bibliography
}
\description{
Imports files using bib format.
}
