% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/read_refs.R
\name{read_refs}
\alias{read_refs}
\title{Import bibliographic data files}
\usage{
read_refs(
  filename,
  tag_naming = "best_guess",
  return_df = TRUE,
  verbose = FALSE,
  locale = vroom::default_locale(),
  ...
)
}
\arguments{
\item{filename}{A path to a filename or vector of filenames containing search
results to import.}

\item{tag_naming}{Either a length-1 character stating how should ris tags be
replaced (see details for a list of options), or an object inheriting from
class \code{tibble} containing user-defined replacement tags.}

\item{return_df}{If \code{TRUE} (default), returns a \code{tibble}; if \code{FALSE},
returns a list.}

\item{verbose}{If \code{TRUE}, prints status updates (defaults to \code{FALSE}).}

\item{locale}{passed to \code{\link[vroom:vroom_lines]{vroom::vroom_lines()}}}

\item{...}{Additional arguments, passed to \code{\link[vroom:vroom]{vroom::vroom()}} or \code{\link[vroom:vroom_lines]{vroom::vroom_lines()}}}
}
\value{
Returns a \code{tibble} unless \code{return_df} is set to \code{FALSE}, when it returns a \code{list}.
}
\description{
Import common bibliographic reference formats  such as \code{.bib}, \code{.ris}, or
\code{.txt}.
}
\details{
Accepted values for \code{tag_naming} are:
'\itemize{
\item \code{"best_guess"}: estimate which database has been used for ris tag replacement,
then fill any gaps with generic tags. Any tags missing from \link{code_lookup} are passed
unchanged.
\item \code{"wos"} Web of Science tags
\item \code{"scopus"} Scopus tags
\item \code{"ovid"} OVID tags
\item \code{"asp"} Academic Search Premier tags
\item \code{"none"} Do not rename tags
\item A \code{tibble} with the following columns: \itemize{
\item \code{"code"} listing the original tags in the source document
\item \code{"field"} listing the replacement column/tag names
\item \code{"order"} (optional) listing the order of columns in the resulting \code{tibble}
}
}
}
\examples{
litsearchr <- c(
  "@article{grames2019,
  title={An automated approach to identifying search terms for
  systematic reviews using keyword co-occurrence networks},
  author={Grames, Eliza M and Stillman, Andrew N and Tingley, Morgan W and Elphick, Chris S},
  journal={Methods in Ecology and Evolution},
  volume={10},
  number={10},
  pages={1645--1654},
  year={2019},
  publisher={Wiley Online Library}
}")

tmp <- tempfile()
writeLines(litsearchr, tmp)
df <- read_refs(tmp, return_df = TRUE, verbose = TRUE)
}
