% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/import_refs.R
\name{remove_factors}
\alias{remove_factors}
\title{Remove factors from an object}
\usage{
remove_factors(z)
}
\arguments{
\item{z}{a data.frame}
}
\value{
a data.frame with all factors converted to character
}
\description{
converts factors to characters to avoid errors with levels
}
