% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/dfm_functions.R
\name{remove_punctuation}
\alias{remove_punctuation}
\title{Remove punctuation from text}
\usage{
remove_punctuation(text, remove_hyphens = FALSE)
}
\arguments{
\item{text}{the text from which to remove punctuation}
}
\value{
the input text with punctuation removed
}
\description{
Removes common punctuation marks from a text
}
