% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/generic_functions.R
\name{remove_stopwords}
\alias{remove_stopwords}
\title{Remove stopwords from text}
\usage{
remove_stopwords(text, language)
}
\arguments{
\item{text}{A character vector containing text from which to remove stopwords.}

\item{language}{A string indicating the language of the text.}
}
\value{
Returns the input text with stopwords removed.
}
\description{
Removes stopwords from text in whichever language is specified.
}
\examples{
remove_stopwords("On the Origin of Species", language="English")
}
