% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/import_refs.R
\name{rollingsum}
\alias{rollingsum}
\title{Computes a rolling sum}
\usage{
rollingsum(a, n = 2L)
}
\arguments{
\item{a}{a character vector}

\item{n}{the number of previous detections to consider}
}
\value{
a number vector
}
\description{
Computes the rolling sum of detections; intended for use in detect_delimiter
}
\examples{
rollingsum(c(1,1,2,3,5))
}
