% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/synthesisr.R
\docType{package}
\name{synthesisr}
\alias{synthesisr}
\title{synthesisr: Import, Assemble, and Deduplicate Bibiliographic Datasets}
\description{
Systematic review searches include multiple databases
 that export results in a variety of formats with overlap in
 coverage between databases. To streamline the process of importing,
 assembling, and deduplicating results, synthesisr recognizes
 bibliographic files exported from databases commonly used for
 systematic reviews and merges results into a standardized format.
}
\section{Functions}{

\strong{Importing and cleaning bibliographic data}
\itemize{
 \item Import bibliographic datasets via \code{\link{read_refs}}, incorporating:
   \itemize{
     \item \code{\link{read_ref}} Import an individual bibliographic dataset
     \item \code{\link{detect_format}} Detects if a file is bib-like or ris-like
     \item \code{\link{detect_delimiter}} Detect delimiter type in ris files
     \item \code{\link{code_lookup}} Tag replacement for ris files
     \item \code{\link{parse_ris}} Parse a vector containing bibliographic data
     \item \code{\link{as.bibliography}} Methods for class 'bibliography'
   }
 \item Clean column and author names via \code{\link{clean_df}}
   \itemize{
     \item \code{\link{clean_names}} Clean column names
     \item \code{\link{clean_authors}} Clean author names
   }
 \item Indexing
   \itemize{
     \item \code{\link{create_index}} Create an ordered string
     \item \code{\link{generate_ids}} Generate unique row IDs from bibliographic data
  }
}

\strong{Data manipulation}
\itemize{
 \item \code{\link{merge_columns}} rbind two data.frames with different numbers of columns
 \item \code{\link{remove_factors}} Remove factors from a data.frame
}

\strong{Formatting}
\itemize{
 \item \code{\link{format_citation}} Return a clean citation from a bibliography or data.frame
 \item \code{\link{add_line_breaks}} Set a maximum character width for strings in a vector
}

\strong{Deduplication}
\itemize{
 \item \code{\link{fuzzdist}} Fuzzy string matching
 \item \code{\link{find_duplicates}} Locate potentially duplicated references
 \item \code{\link{deduplicate}} Return a data.frame with only 'unique' references
}

\strong{Export}
\itemize{
 \item \code{\link{write_refs}} Export bibliographic data. Bundles:
   \itemize{
     \item \code{\link{write_ris}} Export as ris
     \item \code{\link{write_bib}} Export as bib
 }
}
}

