% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/write_bibliography.R
\name{write_bib}
\alias{write_bib}
\title{Write a .bib file}
\usage{
write_bib(x)
}
\arguments{
\item{x}{an object of class bibliography}
}
\value{
a character vector containing references in .bib format
}
\description{
a subfunction of write_bibliography to convert a bibliography to .bib format
}
