% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/write_refs.R
\name{write_refs}
\alias{write_refs}
\title{Export data to a bibliographic format}
\usage{
write_refs(x, format = "ris", write_file = FALSE, filename = NULL)
}
\arguments{
\item{x}{Either a data.frame containing bibliographic information or an object of class bibliography.}

\item{format}{What format should the data be exported as? Options are ris or bib.}

\item{write_file}{If TRUE, saves the result to a text file in the working directory.}

\item{filename}{If write_file is TRUE, the name of the file to be written.}
}
\value{
Returns a character vector containing bibliographic information in the specified format if write_file is FALSE, or saves output to a file if write_file is TRUE.
}
\description{
This function exports data.frames containing bibliographic information to either a .ris or .bib file.
}
\examples{
bibfile <- c(
  "@article{haddaway2019eviatlas,",
  "title={EviAtlas: a tool for visualising evidence synthesis databases},",
  "author={Haddaway, Neal R and Feierman, Andrew and Grainger, Matthew J
  and Gray, Charles T and Tanriver-Ayder, Ezgi and Dhaubanjar, Sanita
  and Westgate, Martin J},",
  "journal={Environmental Evidence},",
  "volume={8},",
  "number={1},",
  "pages={22},",
  "year={2019},",
  "publisher={BioMed Central}",
"}"
)

bib <- read_bib(bibfile)

write_bibliography(bib, format = "bib")
}
