% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/write_refs.R
\name{write_refs}
\alias{write_refs}
\title{Export data to a bibliographic format}
\usage{
write_refs(x, format = "ris", write_file = TRUE, filename = "bibliography")
}
\arguments{
\item{x}{Either a data.frame containing bibliographic information or an object of class bibliography.}

\item{format}{What format should the data be exported as? Options are ris or bib.}

\item{write_file}{If TRUE, saves the result to a text file in the working directory.}

\item{filename}{If write_file is TRUE, the name of the file to be written.}
}
\value{
Returns a character vector containing bibliographic information in the specified format if write_file is FALSE, or saves output to a file if write_file is TRUE.
}
\description{
This function exports data.frames containing bibliographic information to either a .ris or .bib file.
}
\examples{
eviatlas <- c(
  "TY  - JOUR",
  "AU  - Haddaway, Neal R.",
  "AU  - Feierman, Andrew",
  "AU  - Grainger, Matthew J.",
  "AU  - Gray, Charles T.",
  "AU  - Tanriver-Ayder, Ezgi",
  "AU  - Dhaubanjar, Sanita",
  "AU  - Westgate, Martin J.",
  "PY  - 2019",
  "DA  - 2019/06/04",
  "TI  - EviAtlas: a tool for visualising evidence synthesis databases",
  "JO  - Environmental Evidence",
  "SP  - 22",
  "VL  - 8",
  "IS  - 1",
  "SN  - 2047-2382",
  "UR  - https://doi.org/10.1186/s13750-019-0167-1",
  "DO  - 10.1186/s13750-019-0167-1",
  "ID  - Haddaway2019",
  "ER  - "
)

detect_format(eviatlas) # = "parse_ris"
df <- as.data.frame(parse_ris(eviatlas))
write_refs(df, format = "ris")
}
