% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/write_refs.R
\name{write_ris}
\alias{write_ris}
\title{Write a .ris file}
\usage{
write_ris(x)
}
\arguments{
\item{x}{An object of class bibliography.}
}
\value{
Returns a character vector containing references in .ris format.
}
\description{
This is a subfunction of write_bibliography to convert an object of class bibliography to .ris text format.
}
\examples{
file <- c(
  "TY  - JOUR",
  "AU  - Haddaway, Neal R.",
  "AU  - Feierman, Andrew",
  "AU  - Grainger, Matthew J.",
  "AU  - Gray, Charles T.",
  "AU  - Tanriver-Ayder, Ezgi",
  "AU  - Dhaubanjar, Sanita",
  "AU  - Westgate, Martin J.",
  "PY  - 2019",
  "DA  - 2019/06/04",
  "TI  - EviAtlas: a tool for visualising evidence synthesis databases",
  "JO  - Environmental Evidence",
  "SP  - 22",
  "VL  - 8",
  "IS  - 1",
  "SN  - 2047-2382",
  "UR  - https://doi.org/10.1186/s13750-019-0167-1",
  "DO  - 10.1186/s13750-019-0167-1",
  "ID  - Haddaway2019",
  "ER  - "
)

delim <- detect_delimiter(file)

read_medline(prep_ris(file, delim))

ris <- read_ris(prep_ris(file, delim))

write_ris(ris)
}
